pyproject.toml
[tool.poetry]
name = "django-db-anonymiser"
version = "0.3.2"
description = "Django app to create configurable anonymised DB dumps."
authors = [
    "Brendan Smith <brendan.smith@digital.trade.gov.uk>" # /PS-IGNORE
]
readme = "README.md"
packages = [
    { include = "django_db_anonymiser" },
]

[tool.poetry.dependencies]
python = ">3.10.0,<4.0"
boto3 = "^1.40.33"
psycopg2-binary = "^2.9.10"
django-environ = "^0.12.0"
pymysql = "^1.1.2"
django = "^4.2.10"
faker = ">=4.18.0"
pyyaml = "6.0.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.0"


[tool.poetry.group.test.dependencies]
moto = {extras = ["s3"], version = "^5.1.12"}
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
mock = "^5.2.0"
requests = "^2.32.5"
pytest-mock = "^3.15.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
